Better user management

+++ application/libraries/authorization.php
@@ -10,7 +10,8 @@ class Authorization
 {
     public $role;
     public $alias;
-    public $customer_id;
+    public $admin;
+    public $user_id;
 
     /**
      * Check if a request is authorized
@@ -19,49 +20,61 @@ class Authorization
      * @return true if the authorization is successful
      */
     function authorize($rolesToEnforce){
+
+
         $ci =& get_instance();
         if(!$token = $ci->input->get_request_header('Authorization'))
             $this->_throwError('400', ERROR_NO_TOKEN_IN_AUTHORIZATION);
 
         $ci->load->model('public_token');
-        $ci->load->model('admin_token');
+        $ci->load->model('user_token');
+        $ci->load->model('user');
         if(!$dbtoken = $ci->public_token->get_by_token($token))
-            if(!$dbtoken = $ci->admin_token->get_by_token($token))
+            if(!$dbtoken = $ci->user_token->get_by_token($token))
                 $this->_throwError('403', ERROR_INVALID_TOKEN);
 
         $dbtoken = $dbtoken[0];
-        if( $dbtoken->expiration < date('Y-m-d H:i:s', time()) ||
-            $dbtoken->ip != $ci->input->ip_address() ||
-            $dbtoken->user_agent != $ci->input->user_agent())
-        {
-            try{
-                $ci->public_token->delete($dbtoken->id);
-            }catch(ErrorException $e){
-                log_message("Database error: " . $e);
-            }
-            $this->_throwError('403', ERROR_INVALID_TOKEN);
-        }
+        $user = $ci->user->get($dbtoken->user_id);
 
-        if(isset($dbtoken->infoscreen_id)){
-            $infoscreen = $ci->infoscreen->get($dbtoken->infoscreen_id);
-            if(count($infoscreen) == 1)
-                $this->alias = $infoscreen[0]->alias;
-            else
-                $this->_throwError('403', ERROR_INVALID_TOKEN);
-			$this->role = $dbtoken->role;
-        }else{
-            $this->customer_id = $dbtoken->customer_id;
-			$this->role = AUTH_ADMIN;
-        }
-
-        if(!$this->correctRole($this->role, $rolesToEnforce))
-            $this->_throwError('401', ERROR_ROLE);
-
-        // Proper return for tablets
-        if(!empty($this->alias))
-            return $this->alias;
-        else
+        if($user[0]->rights == 1){
+            // Admin user
+            $this->user_id = $dbtoken->user_id;
+            $this->admin = true;
             return true;
+        }else{
+            if( $dbtoken->expiration < date('Y-m-d H:i:s', time()) ||
+                $dbtoken->ip != $ci->input->ip_address() ||
+                $dbtoken->user_agent != $ci->input->user_agent())
+            {
+                try{
+                    $ci->public_token->delete($dbtoken->id);
+                }catch(ErrorException $e){
+                    log_message("Database error: " . $e);
+                }
+                $this->_throwError('403', ERROR_INVALID_TOKEN);
+            }
+
+            if(isset($dbtoken->infoscreen_id)){
+                $infoscreen = $ci->infoscreen->get($dbtoken->infoscreen_id);
+                if(count($infoscreen) == 1)
+                    $this->alias = $infoscreen[0]->alias;
+                else
+                    $this->_throwError('403', ERROR_INVALID_TOKEN);
+    			$this->role = $dbtoken->role;
+            }else{
+                $this->user_id = $dbtoken->user_id;
+    			$this->role = AUTH_ADMIN;
+            }
+
+            if(!$this->correctRole($this->role, $rolesToEnforce))
+                $this->_throwError('401', ERROR_ROLE);
+
+            // Proper return for tablets
+            if(!empty($this->alias))
+                return $this->alias;
+            else
+                return true;
+        }
     }
 
     /**
